+++ objectbox-java/src/main/java/io/objectbox/BoxStore.java
@@ -21,6 +21,7 @@ import org.greenrobot.essentials.collections.LongHashMap;
 import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
+import java.io.PrintStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
@@ -81,6 +82,8 @@ public class BoxStore implements Closeable {
 
     /** The ObjectBox database version this Java library is known to work with. */
     private static final String VERSION = "4.3.1-2025-08-02";
+
+    private static final String OBJECTBOX_PACKAGE_NAME = "objectbox";
     private static BoxStore defaultStore;
 
     /** Currently used DB dirs with values from {@link #getCanonicalPath(File)}. */
@@ -238,6 +241,7 @@ public class BoxStore implements Closeable {
 
     native long nativePanicModeRemoveAllObjects(long store, int entityId);
 
+    private final PrintStream errorOutputStream;
     private final File directory;
     private final String canonicalPath;
     /** Reference to the native store. Should probably get through {@link #getNativeStore()} instead. */
@@ -283,6 +287,7 @@ public class BoxStore implements Closeable {
         relinker = builder.relinker;
         NativeLibraryLoader.ensureLoaded();
 
+        errorOutputStream = builder.errorOutputStream;
         directory = builder.directory;
         canonicalPath = getCanonicalPath(directory);
         verifyNotAlreadyOpen(canonicalPath);
@@ -613,7 +618,7 @@ public class BoxStore implements Closeable {
         // Because write TXs are typically not cached, initialCommitCount is not as relevant than for read TXs.
         int initialCommitCount = commitCount;
         if (debugTxWrite) {
-            System.out.println("Begin TX with commit count " + initialCommitCount);
+            getOutput().println("Begin TX with commit count " + initialCommitCount);
         }
         long nativeTx = nativeBeginTx(getNativeStore());
         if (nativeTx == 0) throw new DbException("Could not create native transaction");
@@ -638,7 +643,7 @@ public class BoxStore implements Closeable {
         // TODO add multithreaded test for this
         int initialCommitCount = commitCount;
         if (debugTxRead) {
-            System.out.println("Begin read TX with commit count " + initialCommitCount);
+            getOutput().println("Begin read TX with commit count " + initialCommitCount);
         }
         long nativeTx = nativeBeginReadTx(getNativeStore());
         if (nativeTx == 0) throw new DbException("Could not create native read transaction");
@@ -693,12 +698,17 @@ public class BoxStore implements Closeable {
                 // (due to all Java APIs doing closed checks).
                 closed = true;
 
+                // Stop accepting new tasks (async calls, query publishers) on the internal thread pool
+                internalThreadPool().shutdown();
+                // Give running tasks some time to finish, print warnings if they do not to help callers fix their code
+                checkThreadTermination();
+
                 List<Transaction> transactionsToClose;
                 synchronized (transactions) {
                     // Give open transactions some time to close (BoxStore.unregisterTransaction() calls notify),
                     // 1000 ms should be long enough for most small operations and short enough to avoid ANRs on Android.
                     if (hasActiveTransaction()) {
-                        System.out.println("Briefly waiting for active transactions before closing the Store...");
+                        getOutput().println("Briefly waiting for active transactions before closing the Store...");
                         try {
                             // It is fine to hold a lock on BoxStore.this as well as BoxStore.unregisterTransaction()
                             // only synchronizes on "transactions".
@@ -708,7 +718,7 @@ public class BoxStore implements Closeable {
                             // If interrupted, continue with releasing native resources
                         }
                         if (hasActiveTransaction()) {
-                            System.err.println("Transactions are still active:"
+                            getErrorOutput().println("Transactions are still active:"
                                     + " ensure that all database operations are finished before closing the Store!");
                         }
                     }
@@ -726,10 +736,6 @@ public class BoxStore implements Closeable {
                 if (handleToDelete != 0) { // failed before native handle was created?
                     nativeDelete(handleToDelete);
                 }
-
-                // When running the full unit test suite, we had 100+ threads before, hope this helps:
-                threadPool.shutdown();
-                checkThreadTermination();
             }
         }
         if (!oldClosedState) {
@@ -740,22 +746,50 @@ public class BoxStore implements Closeable {
         }
     }
 
-    /** dump thread stacks if pool does not terminate promptly. */
+    /**
+     * Waits briefly for the internal {@link #internalThreadPool()} to terminate. If it does not terminate in time,
+     * prints stack traces of the pool threads.
+     */
     private void checkThreadTermination() {
         try {
-            if (!threadPool.awaitTermination(1, TimeUnit.SECONDS)) {
-                int activeCount = Thread.activeCount();
-                System.err.println("Thread pool not terminated in time; printing stack traces...");
-                Thread[] threads = new Thread[activeCount + 2];
+            if (!internalThreadPool().awaitTermination(1, TimeUnit.SECONDS)) {
+                getErrorOutput().println("ObjectBox thread pool not terminated in time." +
+                        " Ensure all async calls have completed and subscriptions are cancelled before closing the Store." +
+                        "\nDumping stack traces of threads on the pool and any using ObjectBox APIs:" +
+                        "\n=== BEGIN OF DUMP ===");
+                // Note: this may not print any pool threads if other threads are started while enumerating
+                // (and the pool threads do not make it into the threads array).
+                Thread[] threads = new Thread[Thread.activeCount()];
                 int count = Thread.enumerate(threads);
                 for (int i = 0; i < count; i++) {
-                    System.err.println("Thread: " + threads[i].getName());
-                    Thread.dumpStack();
+                    Thread thread = threads[i];
+                    if (shouldDumpThreadStackTrace(thread)) {
+                        getErrorOutput().println("Thread: " + thread.getName());
+                        StackTraceElement[] trace = thread.getStackTrace();
+                        for (StackTraceElement traceElement : trace) {
+                            getErrorOutput().println("\tat " + traceElement);
+                        }
+                    }
                 }
+                getErrorOutput().println("=== END OF DUMP ===");
             }
         } catch (InterruptedException e) {
-            e.printStackTrace();
+            e.printStackTrace(getErrorOutput());
         }
+    }
+
+    private boolean shouldDumpThreadStackTrace(Thread thread) {
+        // Dump any threads of the internal thread pool
+        if (thread.getName().startsWith(ObjectBoxThreadPool.THREAD_NAME_PREFIX)) return true;
+
+        // Any other thread might be blocking a thread on the internal pool, so also dump any that appear to use
+        // ObjectBox APIs.
+        StackTraceElement[] trace = thread.getStackTrace();
+        for (StackTraceElement traceElement : trace) {
+            if (traceElement.getClassName().contains(OBJECTBOX_PACKAGE_NAME)) return true;
+        }
+
+        return false;
     }
 
     /**
@@ -894,7 +928,7 @@ public class BoxStore implements Closeable {
         synchronized (txCommitCountLock) {
             commitCount++; // Overflow is OK because we check for equality
             if (debugTxWrite) {
-                System.out.println("TX committed. New commit count: " + commitCount + ", entity types affected: " +
+                getOutput().println("TX committed. New commit count: " + commitCount + ", entity types affected: " +
                         (entityTypeIdsAffected != null ? entityTypeIdsAffected.length : 0));
             }
         }
@@ -1013,10 +1047,10 @@ public class BoxStore implements Closeable {
                 String diagnose = diagnose();
                 String message = attempt + " of " + attempts + " attempts of calling a read TX failed:";
                 if (logAndHeal) {
-                    System.err.println(message);
+                    getErrorOutput().println(message);
                     e.printStackTrace();
-                    System.err.println(diagnose);
-                    System.err.flush();
+                    getErrorOutput().println(diagnose);
+                    getErrorOutput().flush();
 
                     System.gc();
                     System.runFinalization();
@@ -1121,7 +1155,7 @@ public class BoxStore implements Closeable {
      * See also {@link #runInTx(Runnable)}.
      */
     public void runInTxAsync(final Runnable runnable, @Nullable final TxCallback<Void> callback) {
-        threadPool.submit(() -> {
+        internalScheduleThread(() -> {
             try {
                 runInTx(runnable);
                 if (callback != null) {
@@ -1142,7 +1176,7 @@ public class BoxStore implements Closeable {
      * * See also {@link #callInTx(Callable)}.
      */
     public <R> void callInTxAsync(final Callable<R> callable, @Nullable final TxCallback<R> callback) {
-        threadPool.submit(() -> {
+        internalScheduleThread(() -> {
             try {
                 R result = callInTx(callable);
                 if (callback != null) {
@@ -1314,11 +1348,11 @@ public class BoxStore implements Closeable {
 
     @Internal
     public Future<?> internalScheduleThread(Runnable runnable) {
-        return threadPool.submit(runnable);
+        return internalThreadPool().submit(runnable);
     }
 
     @Internal
-    public ExecutorService internalThreadPool() {
+    ExecutorService internalThreadPool() {
         return threadPool;
     }
 
@@ -1335,6 +1369,20 @@ public class BoxStore implements Closeable {
     @Internal
     public TxCallback<?> internalFailedReadTxAttemptCallback() {
         return failedReadTxAttemptCallback;
+    }
+
+    /**
+     * The output stream to print log messages to. Currently {@link System#out}.
+     */
+    private PrintStream getOutput() {
+        return System.out;
+    }
+
+    /**
+     * The error output stream to print log messages to. This is {@link System#err} by default.
+     */
+    private PrintStream getErrorOutput() {
+        return errorOutputStream;
     }
 
     void setDebugFlags(int debugFlags) {